requirements.txt
streamlit
langchain
openai
python-dotenv
ollama
anthropic
mistralai
google-generativeai
google-genai
toml
psutil
requests
pyinstaller